+++ cli/tests/runner.rs
@@ -54,6 +54,7 @@ mod test_interdiff_command;
 mod test_log_command;
 mod test_new_command;
 mod test_next_prev_commands;
+mod test_op_revert_command;
 mod test_operations;
 mod test_parallelize_command;
 mod test_rebase_command;
